Commit programa inventario => normal, reportes 1

+++ database/migrations/2018_08_20_131258_create_detalles_table.php
@@ -18,7 +18,7 @@ class CreateDetallesTable extends Migration
             $table->integer('cantidad');
             $table->float('valor_unitario',11,2);
             $table->float('valor_total',11,2);
-            $table->integer('producto_id')->unsigned();
+            $table->bigInteger('producto_id')->unsigned();
             $table->integer('factura_id')->unsigned();
             $table->timestamps();
