+++ database/migrations/2018_08_30_144811_create_detalle_creditos_table.php
@@ -18,7 +18,7 @@ class CreateDetalleCreditosTable extends Migration
             $table->integer('cantidad');
             $table->float('valor_unitario',11,2);
             $table->float('valor_total',11,2);
-            $table->integer('producto_id')->unsigned();
+            $table->bigInteger('producto_id')->unsigned();
             $table->integer('credito_id')->unsigned();
             $table->timestamps();
