+++ database/migrations/2018_09_13_141729_create_detalles_pedidos_table.php
@@ -17,7 +17,7 @@ class CreateDetallesPedidosTable extends Migration
             $table->increments('id');
             $table->integer('cantidad');
             $table->float('valor_unitario',11,2);
-            $table->integer('producto_id')->unsigned();
+            $table->bigInteger('producto_id')->unsigned();
             $table->integer('pedido_id')->unsigned();
             $table->timestamps();
